app, cmd/gogio: add android foreground permission and service
This adds the permission android.permission.FOREGROUND_SERVICE and adds
GioForegroundService which creates the tray Notification necessary to
implement the Foreground Service.

This adds the method Start to package app, which on android, notifies
the system that the program will perform background work and that it
shouldn't be killed. The foreground service is stopped using the cancel
function returned by Start(). If multiple calls to Start are made, the
foreground service will not be stopped until the final cancel function
has been called.

See https://developer.android.com/guide/components/foreground-services
and https://developer.android.com/training/notify-user/build-notification

Signed-off-by: Masala <masala@riseup.net>